backend/requirements.txt
Flask>=2.3,<4.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-SQLAlchemy>=3.1.0
Flask-SocketIO>=5.3.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
python-socketio>=5.10.0
python-engineio>=4.9.0
gunicorn>=21.0.0
cryptography>=41.0.0
eventlet>=0.35.0
